feat: v2alpha intent writemodel refactoring
